# perf: optimize compute_block_hash_for_seq with unsafe pointer casting