@ryanolson 2025-07-06
Nice. Good spot. I think we can use bytemuck to make this even cleaner.