@xiaguan 2025-07-06
Thanks a lot for your thorough review and suggestions!

I've replaced the unsafe code with `bytemuck`, and added a fallback for big-endian platforms using`cfg`.

I also gave `rayon` a try, but the tasks were too small to really benefit from it—in fact, the overhead ended up making things slower. So I decided to not use it.